Implement GET /todos/{id}
